@@ -2,36 +2,34 @@
 using MediatR;
 using Server.Application.Exceptions;
 
-namespace Server.Application.Behaviors
+namespace Server.Application.Behaviors;
+
+public class ValidationBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse> where TRequest : IRequest<TResponse>
 {
-    public class ValidationBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse> where TRequest : IRequest<TResponse>
+    private readonly IEnumerable<IValidator<TRequest>> _validators;
+
+    public ValidationBehavior(IEnumerable<IValidator<TRequest>> validators)
     {
-        private readonly IEnumerable<IValidator<TRequest>> _validators;
-
-        public ValidationBehavior(IEnumerable<IValidator<TRequest>> validators)
-        {
-            _validators = validators;
-        }
-        public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken, RequestHandlerDelegate<TResponse> next)
-        {
-            if (_validators.Any())
-            {
-                var context = new ValidationContext<TRequest>(request);
-
-                var validationResults = await Task.WhenAll(
-                    _validators.Select(v =>
-                    v.ValidateAsync(context, cancellationToken)));
-
-                var failures = validationResults
-                    .Where(r => r.Errors.Any())
-                    .SelectMany(r => r.Errors)
-                    .ToList();
-
-                if (failures.Any())
-                    throw new ModelValidationException(failures);
-            }
-            return await next();
-        }
+        _validators = validators;
     }
+    public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken, RequestHandlerDelegate<TResponse> next)
+    {
+        if (_validators.Any())
+        {
+            var context = new ValidationContext<TRequest>(request);
 
+            var validationResults = await Task.WhenAll(
+                _validators.Select(v =>
+                v.ValidateAsync(context, cancellationToken)));
+
+            var failures = validationResults
+                .Where(r => r.Errors.Any())
+                .SelectMany(r => r.Errors)
+                .ToList();
+
+            if (failures.Any())
+                throw new ModelValidationException(failures);
+        }
+        return await next();
+    }
 }
