@@ -1,22 +1,16 @@
 ﻿using Microsoft.Extensions.Configuration;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Server.Persistence
+namespace Server.Persistence;
+
+public class Configuration
 {
-    public class Configuration
+     public static T GetSettings<T>(string key)
     {
-         public static T GetSettings<T>(string key)
-        {
-            IConfigurationRoot configuration = new ConfigurationBuilder()
-                .SetBasePath(Directory.GetCurrentDirectory())
-                .AddJsonFile("appsettings.json")
-                .Build();
+        IConfigurationRoot configuration = new ConfigurationBuilder()
+            .SetBasePath(Directory.GetCurrentDirectory())
+            .AddJsonFile("appsettings.json")
+            .Build();
 
-            return configuration.GetSection(key).Get<T>();
-        }
+        return configuration.GetSection(key).Get<T>();
     }
 }
