@@ -1,25 +1,19 @@
 ﻿using FluentValidation.Results;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Server.Application.Exceptions
+namespace Server.Application.Exceptions;
+
+public class ModelValidationException : Exception
 {
-    public class ModelValidationException : Exception
+    public IDictionary<string, string[]> Errors { get; }
+    public ModelValidationException() : base("One or more validation failures have occured.")
     {
-        public IDictionary<string, string[]> Errors { get; }
-        public ModelValidationException() : base("One or more validation failures have occured.")
-        {
-            Errors = new Dictionary<string, string[]>();
-        }
+        Errors = new Dictionary<string, string[]>();
+    }
 
-        public ModelValidationException(IEnumerable<ValidationFailure> failures) : this()
-        {
-            Errors = failures
-            .GroupBy(e => e.PropertyName, e => e.ErrorMessage)
-            .ToDictionary(failureGroup => failureGroup.Key, failureGroup => failureGroup.ToArray());
-        }
+    public ModelValidationException(IEnumerable<ValidationFailure> failures) : this()
+    {
+        Errors = failures
+        .GroupBy(e => e.PropertyName, e => e.ErrorMessage)
+        .ToDictionary(failureGroup => failureGroup.Key, failureGroup => failureGroup.ToArray());
     }
 }
