@@ -3,16 +3,15 @@ using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using System.Reflection;
 
-namespace Server.Application
+namespace Server.Application;
+
+public static class ServiceRegistration
 {
-    public static class ServiceRegistration
+    public static void AddApplicationRegistration(this IServiceCollection services)
     {
-        public static void AddApplicationRegistration(this IServiceCollection services)
-        {
-            var assm = Assembly.GetExecutingAssembly();
-            services.AddValidatorsFromAssembly(assm);
-            services.AddAutoMapper(assm);
-            services.AddMediatR(assm);
-        }
+        var assm = Assembly.GetExecutingAssembly();
+        services.AddValidatorsFromAssembly(assm);
+        services.AddAutoMapper(assm);
+        services.AddMediatR(assm);
     }
 }
