@@ -1,24 +1,18 @@
 ﻿using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.Configuration;
-using System;
-using System.Collections.Generic;
 using System.Data;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Server.Persistence.Context
+namespace Server.Persistence.Context;
+
+public class DapperContext
 {
-    public class DapperContext
+    private readonly IConfiguration _configuration;
+    private readonly string _connectionString;
+    public DapperContext(IConfiguration configuration)
     {
-        private readonly IConfiguration _configuration;
-        private readonly string _connectionString;
-        public DapperContext(IConfiguration configuration)
-        {
-            _configuration = configuration;
-            _connectionString = Configuration.GetSettings<string>("ConnectionStrings:ConnectionString");
-        }
-        public IDbConnection CreateConnection()
-            => new SqlConnection(_connectionString);
+        _configuration = configuration;
+        _connectionString = Configuration.GetSettings<string>("ConnectionStrings:ConnectionString");
     }
+    public IDbConnection CreateConnection()
+        => new SqlConnection(_connectionString);
 }
