@@ -1,35 +1,27 @@
 ﻿using FluentMigrator.Runner;
-using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Server.Persistence.Extensions
+namespace Server.Persistence.Extensions;
+
+public static class MigrationManager
 {
-    public static class MigrationManager
+    public static IHost MigrateDatabase(this IHost host)
     {
-        public static IHost MigrateDatabase(this IHost host)
+        using (var scope = host.Services.CreateScope())
         {
-            using (var scope = host.Services.CreateScope())
-            {
-                var migrationService = scope.ServiceProvider.GetRequiredService<IMigrationRunner>();
+            var migrationService = scope.ServiceProvider.GetRequiredService<IMigrationRunner>();
 
-                try
-                {
-                    migrationService.ListMigrations();
-                    migrationService.MigrateUp();
-                }
-                catch
-                {
-                    throw;
-                }
+            try
+            {
+                migrationService.ListMigrations();
+                migrationService.MigrateUp();
             }
-            return host;
+            catch
+            {
+                throw;
+            }
         }
+        return host;
     }
 }
