@@ -1,15 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Server.Application.Dto;
 
-namespace Server.Application.Dto
+public class UserDto
 {
-    public class UserDto
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; } = null!;
-        public string Image { get; set; } = null!;
-    }
+    public Guid Id { get; set; }
+    public string Name { get; set; } = null!;
+    public string Image { get; set; } = null!;
 }
