@@ -1,18 +1,10 @@
-﻿using Server.Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Server.Application.Interfaces.Repository;
 
-namespace Server.Application.Interfaces.Repository
+public interface IGenericRepository<T> where T : class
 {
-    public interface IGenericRepository<T> where T : class
-    {
-        Task<T> GetById(Guid id);
-        Task<bool> Create(T entity);
-        Task<bool> Update(T entity);
-        Task<bool> Delete(T entity);
-        Task<List<T>> List();
-    }
+    Task<T> GetById(Guid id);
+    Task<bool> Create(T entity);
+    Task<bool> Update(T entity);
+    Task<bool> Delete(T entity);
+    Task<List<T>> List();
 }
