@@ -1,15 +1,8 @@
-﻿using Server.Application.Dto;
-using Server.Application.Features.User.Commands;
-using Server.Application.Features.User.Queries;
-using Server.Application.Handlers.User.Commands;
-using Server.Application.Wrappers;
-using Server.Domain.Entities;
+﻿using Server.Domain.Entities;
 
+namespace Server.Application.Interfaces.Repository;
 
-namespace Server.Application.Interfaces.Repository
+public interface IUserRepository: IGenericRepository<User>
 {
-    public interface IUserRepository: IGenericRepository<User>
-    {
-        Task<User> GetByEmail(string email);
-    }
+    Task<User> GetByEmail(string email);
 }
