@@ -1,21 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Server.Application.Password;
 
-namespace Server.Application.Password
+public static class Encryption
 {
-    public static class Encryption
+    public static string EncryptPassword(string pPassword)
     {
-        public static string EncryptPassword(string pPassword)
-        {
-            return BCrypt.Net.BCrypt.HashPassword(pPassword);
-        }
+        return BCrypt.Net.BCrypt.HashPassword(pPassword);
+    }
 
-        public static bool VerifyPassword(string pPassword, string pHash)
-        {
-            return BCrypt.Net.BCrypt.Verify(pPassword, pHash);
-        }
+    public static bool VerifyPassword(string pPassword, string pHash)
+    {
+        return BCrypt.Net.BCrypt.Verify(pPassword, pHash);
     }
 }
