@@ -1,16 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Server.Domain.Entities;
 
-namespace Server.Domain.Entities
+public class Settings
 {
-    public class Settings
-    {
-        public string Secret { get; set; } = null!;
-        public string ValidIssuer { get; set; } = null!;
-        public string ValidAudience { get; set; } = null!;
-        public int ValidityPeriod { get; set; }
-    }
+    public string Secret { get; set; } = null!;
+    public string ValidIssuer { get; set; } = null!;
+    public string ValidAudience { get; set; } = null!;
+    public int ValidityPeriod { get; set; }
 }
