@@ -1,24 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Server.Persistence.Helpers;
 
-namespace Server.Persistence.Helpers
+public class RepositoryHelper
 {
-    public class RepositoryHelper
+    public static string GetTableName<T>()
     {
-        public static string GetTableName<T>()
-        {
-            var entityType = typeof(T);
-            return entityType.Name + "s";
-        }
+        var entityType = typeof(T);
+        return entityType.Name + "s";
+    }
 
-        public static IEnumerable<string> GetPropertyNames<T>()
-        {
-            return typeof(T).GetProperties().Select(p => p.Name);
-        }
+    public static IEnumerable<string> GetPropertyNames<T>()
+    {
+        return typeof(T).GetProperties().Select(p => p.Name);
     }
 }
